new api endpoint for navbar, product handling logic changes, some api changes as well

+++ src/main/java/alledrogo/controller/AuthenticationController.java
@@ -1,12 +1,15 @@
 package alledrogo.controller;
 
 import alledrogo.data.entity.UserEntity;
+import alledrogo.data.enums.UserRole;
+import alledrogo.data.enums.UserStatus;
 import alledrogo.io.request.AuthenticationRequest;
 import alledrogo.io.response.MessageResponse;
 import alledrogo.security.jwt.JwtUtils;
 import alledrogo.service.UserService;
 
 import jakarta.validation.Valid;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -55,9 +58,7 @@ public class AuthenticationController {
 
         HttpHeaders headers = new HttpHeaders();
 
-        String expirationTimeString = getExpirationTimeString();
-
-        headers.add("Set-Cookie", "token=" + token + "; Path=/; HttpOnly; Secure; SameSite=None; Expires=" + expirationTimeString);
+        headers.add("Set-Cookie", "token=" + token + "; Path=/; HttpOnly; Secure; SameSite=None; Expires=" + getExpirationTimeString());
 
         return ResponseEntity.ok()
                 .headers(headers)
@@ -70,9 +71,7 @@ public class AuthenticationController {
 
         HttpHeaders headers = new HttpHeaders();
 
-        String expirationTimeString = expireTokens();
-
-        headers.add("Set-Cookie", "token=" + "; Path=/; HttpOnly; Secure; SameSite=None; Expires=" + expirationTimeString);
+        headers.add("Set-Cookie", "token=" + "; Path=/; HttpOnly; Secure; SameSite=None; Expires=" + expireTokens());
 
         return ResponseEntity.ok()
                 .headers(headers)
@@ -84,12 +83,26 @@ public class AuthenticationController {
     public ResponseEntity<?> signupUser(@Valid @RequestBody AuthenticationRequest authenticationRequest) {
 
         String username = authenticationRequest.getUsername();
+        String email = authenticationRequest.getEmail();
 
-        if (userService.existsByUsername(username)) {
-            return ResponseEntity.badRequest().body(new MessageResponse("Username is already taken!"));
+        if (StringUtils.isEmpty(username) || StringUtils.isEmpty(authenticationRequest.getPassword()) || StringUtils.isEmpty(email)) {
+            return ResponseEntity.badRequest().body(new MessageResponse("All fields are required"));
         }
 
-        UserEntity userEntity = new UserEntity(username, passwordEncoder.encode(authenticationRequest.getPassword()));
+        if (userService.existsByUsername(username)) {
+            return ResponseEntity.badRequest().body(new MessageResponse("An account with the given Username already exists"));
+        }
+
+        if (userService.existsByEmail(email)) {
+            return ResponseEntity.badRequest().body(new MessageResponse("An account with the given Email address already exists!"));
+        }
+
+        UserEntity userEntity = new UserEntity(username,
+                passwordEncoder.encode(authenticationRequest.getPassword()),
+                UserRole.ROLE_USER,
+                UserStatus.STATUS_ACTIVE,
+                email);
+
         boolean createUser = userService.createUser(userEntity);
 
         if (createUser){
@@ -103,9 +116,7 @@ public class AuthenticationController {
 
             HttpHeaders headers = new HttpHeaders();
 
-            String expirationTimeString = getExpirationTimeString();
-
-            headers.add("Set-Cookie", "token=" + token + "; Path=/; HttpOnly; Secure; SameSite=None; Expires=" + expirationTimeString);
+            headers.add("Set-Cookie", "token=" + token + "; Path=/; HttpOnly; Secure; SameSite=None; Expires=" + getExpirationTimeString());
 
             return ResponseEntity.ok()
                     .headers(headers)
@@ -122,12 +133,7 @@ public class AuthenticationController {
 
             HttpHeaders headers = new HttpHeaders();
 
-            String expirationTimeString = getExpirationTimeString();
-
-            headers.add("Set-Cookie", "userId=" + userEntity.getId() + "; Path=/; Secure; SameSite=None; Expires=" + expirationTimeString);
-            headers.add("Set-Cookie", "username=" + userEntity.getUsername() + "; Path=/; Secure; SameSite=None; Expires=" + expirationTimeString);
-            headers.add("Set-Cookie", "role=" + userEntity.getRole() + "; Path=/; Secure; SameSite=None; Expires=" + expirationTimeString);
-            headers.add("Set-Cookie", "balance=" + userEntity.getBalance() + "; Path=/; Secure; SameSite=None; Expires=" + expirationTimeString);
+            headers.add("Set-Cookie", "username=" + userEntity.getUsername() + "; Path=/; Secure; SameSite=None; Expires=" + getExpirationTimeString());
 
             return ResponseEntity.ok()
                     .headers(headers)
@@ -138,16 +144,21 @@ public class AuthenticationController {
     }
 
     @DeleteMapping("/{id}/delete")
-    public ResponseEntity<?> deleteUser(@Valid @RequestBody AuthenticationRequest authenticationRequest) {
+    public ResponseEntity<?> deleteUser(@Valid @RequestBody AuthenticationRequest authenticationRequest, @CookieValue(name = "token") String token) {
 
-        String username = authenticationRequest.getUsername();
+        String requestUsername = authenticationRequest.getUsername();
+        String tokenUsername = jwtUtils.getUserNameFromJwtToken(token);
+
+        if (!requestUsername.equals(tokenUsername)) {
+            return ResponseEntity.badRequest().body(new MessageResponse("Something went wrong!."));
+        }
 
         Authentication authentication = authenticationManager
                 .authenticate(new UsernamePasswordAuthenticationToken(authenticationRequest.getUsername(), authenticationRequest.getPassword()));
 
         SecurityContextHolder.getContext().setAuthentication(authentication);
 
-        UserEntity userEntity = userService.findByUsername(username);
+        UserEntity userEntity = userService.findByUsername(requestUsername);
 
         boolean deleteUser = userService.deleteUser(userEntity);
         if (deleteUser){
@@ -158,16 +169,21 @@ public class AuthenticationController {
     }
 
     @PutMapping("/{id}/update")
-    public ResponseEntity<?> updateUser(@Valid @RequestBody AuthenticationRequest authenticationRequest) {
+    public ResponseEntity<?> updateUser(@Valid @RequestBody AuthenticationRequest authenticationRequest,  @CookieValue(name = "token") String token) {
 
-        String username = authenticationRequest.getUsername();
+        String requestUsername = authenticationRequest.getUsername();
+        String tokenUsername = jwtUtils.getUserNameFromJwtToken(token);
+
+        if (!requestUsername.equals(tokenUsername)) {
+            return ResponseEntity.badRequest().body(new MessageResponse("Something went wrong!."));
+        }
 
         Authentication authentication = authenticationManager
-                .authenticate(new UsernamePasswordAuthenticationToken(username, authenticationRequest.getPassword()));
+                .authenticate(new UsernamePasswordAuthenticationToken(requestUsername, authenticationRequest.getPassword()));
 
         SecurityContextHolder.getContext().setAuthentication(authentication);
 
-        UserEntity userEntity = userService.findByUsername(username);
+        UserEntity userEntity = userService.findByUsername(requestUsername);
 
         //TODO updating user
